# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BatchDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  dto.OperateByID:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  dto.PageResult:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
  dto.SearchWithPage:
    properties:
      info:
        type: string
      order:
        type: string
      orderBy:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
    required:
    - page
    - pageSize
    type: object
...
  dto.SnapshotCreate:
    properties:
      description:
        maxLength: 256
        type: string
      from:
        enum:
        - OSS
        - S3
        - SFTP
        - MINIO
        - COS
        - KODO
        - OneDrive
        type: string
      id:
        type: integer
    required:
    - from
    type: object
  dto.SnapshotImport:
    properties:
      description:
        maxLength: 256
        type: string
      from:
        type: string
      names:
        items:
          type: string
        type: array
    type: object
  dto.SnapshotRecover:
    properties:
      id:
        type: integer
      isNew:
        type: boolean
      reDownload:
        type: boolean
    required:
    - id
    type: object
...
  dto.UpdateDescription:
    properties:
      description:
        maxLength: 256
        type: string
      id:
        type: integer
    required:
    - id
    type: object
...
host: localhost
info:
  contact: {}
  description: 开源Linux面板
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 1Panel
  version: "1.0"
paths:
...
  /settings/snapshot:
    post:
      consumes:
      - application/json
      description: 创建系统快照
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SnapshotCreate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create system snapshot
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - from
        - description
        formatEN: Create system backup [description] to [from]
        formatZH: 创建系统快照 [description] 到 [from]
        paramKeys: []
  /settings/snapshot/del:
    post:
      consumes:
      - application/json
      description: 删除系统快照
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchDeleteReq'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete system backup
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions:
        - db: snapshots
          input_column: id
          input_value: ids
          isList: true
          output_column: name
          output_value: name
        bodyKeys:
        - ids
        formatEN: Delete system backup [name]
        formatZH: 删除系统快照 [name]
        paramKeys: []
  /settings/snapshot/description/update:
    post:
      consumes:
      - application/json
      description: 更新快照描述信息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDescription'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update snapshot description
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions:
        - db: snapshots
          input_column: id
          input_value: id
          isList: false
          output_column: name
          output_value: name
        bodyKeys:
        - id
        - description
        formatEN: The description of the snapshot [name] is modified => [description]
        formatZH: 快照 [name] 描述信息修改 [description]
        paramKeys: []
  /settings/snapshot/import:
    post:
      consumes:
      - application/json
      description: 导入已有快照
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SnapshotImport'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Import system snapshot
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - from
        - names
        formatEN: Sync system snapshots [names] from [from]
        formatZH: 从 [from] 同步系统快照 [names]
        paramKeys: []
  /settings/snapshot/recover:
    post:
      consumes:
      - application/json
      description: 从系统快照恢复
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SnapshotRecover'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Recover system backup
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions:
        - db: snapshots
          input_column: id
          input_value: id
          isList: false
          output_column: name
          output_value: name
        bodyKeys:
        - id
        formatEN: Recover from system backup [name]
        formatZH: 从系统快照 [name] 恢复
        paramKeys: []
  /settings/snapshot/rollback:
    post:
      consumes:
      - application/json
      description: 从系统快照回滚
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SnapshotRecover'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Rollback system backup
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions:
        - db: snapshots
          input_column: id
          input_value: id
          isList: false
          output_column: name
          output_value: name
        bodyKeys:
        - id
        formatEN: Rollback from system backup [name]
        formatZH: 从系统快照 [name] 回滚
        paramKeys: []
  /settings/snapshot/search:
    post:
      consumes:
      - application/json
      description: 获取系统快照列表分页
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SearchWithPage'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PageResult'
      security:
      - ApiKeyAuth: []
      summary: Page system snapshot
      tags:
      - System Setting
  /settings/snapshot/status:
    post:
      consumes:
      - application/json
      description: 获取快照状态
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OperateByID'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load Snapshot status
      tags:
      - System Setting
...
swagger: "2.0"